conflicts: move `Merge` tests to `merge` module
I missed the tests when I moved the type.
